backend/requirements.txt
# Web framework
Flask==3.1.2
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.3.0
itsdangerous==2.2.0
blinker==1.9.0

# Database
SQLAlchemy==2.0.43
Flask-SQLAlchemy==3.1.1
alembic==1.16.5
Flask-Migrate==4.1.0
greenlet==3.2.4
Mako==1.3.10

# Authentication (NUEVAS - hay que instalarlas)
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2

# CORS
flask-cors==6.0.1
flask-talisman==1.1.0

# Deployment
gunicorn==25.1.0

# Environment variables
python-dotenv==1.1.1

# Image processing (NUEVAS - opcionales para fotos)
Pillow==10.2.0
cloudinary==1.36.0  # Opcional, para subir imágenes a la nube

# Testing
pytest==8.0.0

# Utils
typing_extensions==4.15.0
colorama==0.4.6
MarkupSafe==3.0.3